+++ src/text/str/ext_str.rs
@@ -137,7 +137,7 @@ impl ExtStr for str {
             let mut separator_turn = true; // start writing the separator
 
             let mut num_buf = AsciiDigits(num).digits();
-            let mut num_bytes = Slice::trim_leading_bytes(&num_buf, b'0');
+            let mut num_bytes = Slice::trim_leading(&num_buf, b'0');
             // IMPROVE:BENCH use NumToStr
             // let mut num_buf = [0u8; 22];
             // let mut num_bytes = num.to_bytes_base(10, &mut num_buf);
@@ -154,7 +154,7 @@ impl ExtStr for str {
                     num = index;
 
                     num_buf = AsciiDigits(num).digits();
-                    num_bytes = Slice::trim_leading_bytes(&num_buf, b'0');
+                    num_bytes = Slice::trim_leading(&num_buf, b'0');
                     // IMPROVE: use NumToStr
                     // num_bytes = num.to_bytes_base(10, &mut num_buf);